requirements.txt
streamlit
pymongo
google-generativeai
requests
Pillow